Add example mirroring 5_20.rb
Demonstrates a few of the ways the code might be indicating the presence
of an undiscovered duck.

Checking for a specific constructor; checking if the argument is an
instance of a class; checking if a property is 'in' the argument object.
All are asking to replaced by a duck, if you can identify the common
message that could be sent to each.