Add example mirroring 6_40.rb
Meant to warn about the dangers of forgetting to send super from
subclasses. JavaScript won't allow you to reference 'this' before
calling 'super' in the constructor.

The requirement to call 'super' in the spares method, on the other hand,
will not be enforced by the language. Forgetting to make this call is
quite possible.